+++ templates/empty/blakron.config.ts
@@ -2,12 +2,12 @@ export default {
 	target: 'html5',
 	entry: 'src/Main.ts',
 	output: { dir: 'bin-debug' },
+	html: { template: 'template/web/index.html' },
 	stage: {
 		width: 640,
 		height: 1136,
 		scaleMode: 'showAll',
 		orientation: 'auto',
 		frameRate: 60,
-		background: '#000000',
 	},
 };